pyproject.toml
[project]
name = "IKB_MCP_Server"
version = "0.1.0"
description = "A free, plug-and-play knowledge base. Built-in with 10,000+ high-quality insight reports, packaged as MCP Server, and secure local data storage."
authors = [
    {name = "Shawn Hui", email = "shawnhui1012@gmail.com"}
]
license = "MIT"
keywords = ["mcp", "konwledge base", "plug-and-play", "insights"]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.9.3",
    "openai>=1.84.0",
    "pillow>=11.2.1",
    "pymupdf>=1.26.0",
    "python-dotenv>=1.1.0",
    "tinydb>=4.8.2",
    "tqdm>=4.67.1",
]
